+++ packages/review-editor/components/PRSelector.tsx
@@ -2,7 +2,7 @@ import React, { useState, useCallback, useEffect, useMemo } from 'react';
 import { SearchableSelect } from '@plannotator/ui/components/SearchableSelect';
 import { PullRequestIcon } from '@plannotator/ui/components/PullRequestIcon';
 import { getItem, setItem } from '@plannotator/ui/utils/storage';
-import type { PRListItem } from '@plannotator/shared/pr-provider';
+import type { PRListItem } from '@plannotator/shared/pr-types';
 
 type PRItem = PRListItem;
