+++ packages/review-editor/hooks/usePRSession.ts
@@ -1,5 +1,5 @@
 import { useState, useCallback } from 'react';
-import type { PRMetadata } from '@plannotator/shared/pr-provider';
+import type { PRMetadata } from '@plannotator/shared/pr-types';
 import type { PRDiffScope, PRDiffScopeOption, PRStackInfo, PRStackTree } from '@plannotator/shared/pr-stack';
 
 export interface PRSessionState {
